Refactor page creation
